% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcHyp.R
\name{calcHyp}
\alias{calcHyp}
\title{Calculate hypotenuse}
\usage{
calcHyp(leg1, leg2)
}
\arguments{
\item{leg1}{the length of one leg of the triangle}

\item{leg2}{the length of the other leg of the triangle}
}
\value{
hyp the length of the hypotenuse
}
\description{
Given two lengths of the legs of a right triangle,
the function returns the length of the hypotenuse
}
\examples{
calcHyp(3, 4)
calcHyp(5, 12)
}

