% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FindHypotenuse.R
\docType{package}
\name{FindHypotenuse}
\alias{FindHypotenuse}
\alias{FindHypotenuse-package}
\title{A toy package}
\description{
This package is boring.
It doesn't do much.
}
\details{
We can put more here if we want.
}

